# Conflicts:
#	main.cpp

+++ main.cpp
@@ -13,8 +13,7 @@
 #include "Include/scene.h"
 #include <iostream>
 #include <string>
-#include <chrono>
-#include <iomanip>
+#include <omp.h>
 
 int main(int argc, char** argv) {
     // Read command line parameters to get scene file
@@ -30,16 +29,16 @@ int main(int argc, char** argv) {
     const Scene scene = parseSceneFile(sceneFileName, img_width, img_height, imgName);
 
     Image outputImg = Image(img_width, img_height);
-    float imgW = img_width, imgH = img_height;
-    float halfW = imgW / 2, halfH = imgH / 2;
-    float d = halfH / tanf(scene.camera_fov_ha * (M_PI / 180.0f));
+    double imgW = img_width, imgH = img_height;
+    double halfW = imgW / 2, halfH = imgH / 2;
+    double d = halfH / tanf(scene.camera_fov_ha * (M_PI / 180.0));
 
     auto t_total_start = std::chrono::steady_clock::now();
-    
+
     for (int i = 0; i < img_width; i++) {
         for (int j = 0; j < img_height; j++) {
-            float u = halfW - i + 0.5;
-            float v = halfH - j + 0.5;
+            double u = halfW - i + 0.5;
+            double v = halfH - j + 0.5;
             Point3 p = scene.camera_pos - d * scene.camera_fwd + u * scene.camera_right + v * scene.camera_up;
             Ray ray(scene.camera_pos, p - scene.camera_pos);
             Color result = rayTrace(ray, scene.max_depth, scene);
@@ -72,4 +71,6 @@ int main(int argc, char** argv) {
 
     return 0;
 }
+
+
